requirements.txt
alabaster==0.7.9
Babel==2.3.4
docutils==0.12
imagesize==0.7.1
Jinja2==2.8
MarkupSafe==0.23
py==1.4.31
Pygments==2.1.3
pytest==3.0.2
pytest-sugar==0.7.1
pytz==2016.7
PyYAML==3.12
requests==2.11.1
requests-mock==1.1.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.8
sphinx-rtd-theme==0.1.9
termcolor==1.1.0
